Rename to jsx and add new ui code
